requirements.txt
streamlit>=1.36.0
requests>=2.31.0
pillow>=10.3.0
supabase>=2.6.0
python-dateutil>=2.9.0
